feat: causes signup form initial work
* Added sign up form and components, couple of minor fixes and organization of templates

GD-162
