+++ src/pages/Causes/Causes.vue
@@ -1,7 +1,11 @@
+<!-- TODO
+main-class="kv-tailwind" can't be added to www-page here because kv-tailwind reset classes are overwriting
+BT drop in css supplied by braintree in src/components/Causes/CausesDropInPaymentWrapper.vue
+Add main-class="kv-tailwind" once BT drop in and tailwind issues are worked out.
+And remove kv-tailwind from child views-->
 <template>
 	<www-page
 		:header-theme="headerTheme"
-		main-class="kv-tailwind"
 	>
 		<router-view />
 	</www-page>
